Moved tests to LocalDateTimeE2eTest
